requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
cdifflib==1.2.5
chardet==4.0.0
discord.py==1.7.3
idna==3.3
multidict==6.0.2
numpy==1.22.4
pandas==1.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
typing-extensions==4.2.0
yarl==1.7.2
